# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Learning Platform API Documentation",
        "description": "Backend API endpoints for React Web and Flutter Mobile application.",
        "contact": {
            "email": "admin@yourdomain.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://dev-api.yourdomain.com",
            "description": "Development Server"
        },
        {
            "url": "https://api.yourdomain.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/v1/checkout/init": {
            "post": {
                "tags": [
                    "Checkout & Enrollment"
                ],
                "summary": "Initialize UddoktaPay Checkout",
                "description": "Calculates the final price securely and returns the UddoktaPay payment URL. If the course is 100% free (via coupon or base price), it enrolls the user instantly.",
                "operationId": "b25513a3f79c7b58084daf284e3f54aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_id"
                                ],
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "coupon_code": {
                                        "type": "string",
                                        "example": "REACT50"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Checkout initialized successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "is_free": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "payment_url": {
                                                    "type": "string",
                                                    "example": "https://pay.uddoktapay.com/checkout/12345"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid coupon or already enrolled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token to access secured endpoints",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}